web/requirements.txt
psycopg2==2.7.7
redis==3.2.1
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Pillow==6.2.1
cssselect2==0.2.2
defusedxml==0.6.0
Django==3.0
django-file-resubmit==0.5.2
django-sendfile==0.3.11
django-sendfile2==0.5.1
django-tinymce==2.8.0
html5lib==1.0.1
idna==2.8
matching==1.1
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tinycss2==1.0.2
urllib3==1.25.8
webencodings==0.5.1
gunicorn==20.0.4